Componetize monolithic app (#2)
* Create Basket component

Created a separate component to render the basket.  Takes a single prop count to indicate the
number of items currently in the basket.

* Create BuyButton component

Created a separate component to handle the buy button.  Takes a callback method for click handling and
a price to render into the button.

* Create Recommendations component

Refactored recommendations section out to separate component with unique functionality.