Create first MFE application, team-inspire
Using the create-mf-app npx script, initialized a new React application for the recommendation engine / components.
This will be the first aspect of the site that will exist as a wholly independent entity.  No cross-dependencies will
exist between the primary application and this component.  The webpack.config.js file for this new application includes
the ModuleFederation plugin specification, the first instace so far, detailing how to expose one or more components
for consumption (remotely) by another application.

Added a new row to the branch table describing the purpose of this branch as well.