Introducing team inspire (#3)
* Create first MFE application, team-inspire

Using the create-mf-app npx script, initialized a new React application for the recommendation engine / components.
This will be the first aspect of the site that will exist as a wholly independent entity.  No cross-dependencies will
exist between the primary application and this component.  The webpack.config.js file for this new application includes
the ModuleFederation plugin specification, the first instace so far, detailing how to expose one or more components
for consumption (remotely) by another application.

Added a new row to the branch table describing the purpose of this branch as well.

* Implement recommendations component

Create Recommendations component implementation in the team-inspire project.  Add the component to the
App component of this project to ensure it can be rendered when in use.  This will demonstrate hosting
the component in its own space.

* Update webpack to expose recommendations

Added an entry to the exposes portion of the ModuleFederation plugin to expose the recommendations component.
This _essentially_ makes the component a micro-frontend.  Next, we'll refer to this component in the main
application to pull the MFE into the main app and display it.

* Add webpack configuration to host app

Updated webpack.config.js to include a Module Federation plugin spec for bringing in the recommendations
MFE from the new app.  Importing module as team_inspire.

* Refactor App component to use MFE

Refactor usage of Recommendations component to lazily import remote MFE component and consume it on the page.
This replaces the direct usage of the in-app Recommendations component.

* Remove the recommendation component

Removed the Recommendations component from the main monolithic application in favor of the MFE component
being packaged and provided by webpack from the team-inspire project.