Initialize an Nx workspace with monolith app
This branch/commit introduces the Nx Workspace and build tools as a model
for managing micro frontends.  The monolith app is recreated as a react
app, iniitalized as a module federation host application with no remotes.

This is the framework on which we will progress breaking the app up using
Nx tools to assist in the MFE process.  Not that very little about this
method is different than the manual method we used to create MFE
applications using the `npx create-mf-app` tool and multiple folders.

For a simple application such as this one, the benefits of Nx are not
readily easy to see, the refactoring tools, build assistance and single
node_modules folder management help reduce complexity of the on-machine
development, beyond that, the in-built generators provided by Nx make
many manual and repeptitive tasks much simpler.  It is highly recommended
that the reader visit the [Nx Website](https://nx.dev) to learn more
about the toolset.